requirements.txt
python-dotenv
groq
python-docx
langchain
sentence-transformers
scikit-learn
pandas
langchain_community
